Conflicts:
	CMakeLists.txt
	include/clover_field_order.h
	include/color_spinor_field_order.h
	include/cub_helper.cuh
	include/gauge_field_order.h
	include/kernels/block_orthogonalize.cuh
	include/kernels/coarse_op_preconditioned.cuh
	include/quda_cuda_api.h
	include/texture.h
	include/thrust_helper.cuh
	lib/blas_cublas.cu
	lib/coarse_op.cuh
	lib/coarse_op_preconditioned.cu
	lib/color_spinor_pack.cu
	lib/cuda_color_spinor_field.cpp
	lib/cuda_gauge_field.cpp
	lib/dslash_coarse.cu
	lib/extract_gauge_ghost_helper.cuh
	lib/gauge_fix_fft.cu
	lib/gauge_fix_ovr.cu
	lib/pgauge_exchange.cu
	lib/pgauge_heatbath.cu
	lib/pgauge_init.cu
	lib/quda_cuda_api.cpp
	tests/CMakeLists.txt
	tests/staggered_invert_test.cpp

+++ include/blas_cublas.h
@@ -25,7 +25,8 @@ namespace quda {
        @param[in] Location of the input/output data
        @return Number of flops done in this computation
     */
-    long long BatchInvertMatrix(void *Ainv, void* A, const int n, const int batch, QudaPrecision precision, QudaFieldLocation location);
+    long long BatchInvertMatrix(void *Ainv, void *A, const int n, const uint64_t batch, QudaPrecision precision,
+                                QudaFieldLocation location);
 
   } // namespace cublas
